requirements.txt
alive_progress
sqlalchemy
pymysql
beautifulsoup4
requests
pydantic
python-dotenv